+++ android/app/src/direct/kotlin/app/deterministic/todo/deterministic_todo/MainActivity.kt
@@ -1,5 +1,11 @@
 package app.deterministic.todo.deterministic_todo
 
 import io.flutter.embedding.android.FlutterActivity
+import io.flutter.embedding.engine.FlutterEngine
 
-class MainActivity : FlutterActivity()
+class MainActivity : FlutterActivity() {
+    override fun configureFlutterEngine(flutterEngine: FlutterEngine) {
+        super.configureFlutterEngine(flutterEngine)
+        RunTrackerChannel.register(this, flutterEngine)
+    }
+}