+++ android/app/src/play/kotlin/app/deterministic/todo/deterministic_todo/MainActivity.kt
@@ -22,6 +22,7 @@ class MainActivity : FlutterActivity() {
 
     override fun configureFlutterEngine(flutterEngine: FlutterEngine) {
         super.configureFlutterEngine(flutterEngine)
+        RunTrackerChannel.register(this, flutterEngine)
         updateManager = AppUpdateManagerFactory.create(this)
         updateManager.registerListener(installListener)
         MethodChannel(